+++ audio/file_extension.py
@@ -2,3 +2,4 @@ from enum import Enum
 
 class FileExtension(Enum):
     MP3 = '.mp3'
+    M4A = '.m4a'